+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/controller/AuthController.kt
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.MemberResponse
-import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.SigninRequest
-import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.SigninResponse
-import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.SignupRequest
+import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.response.MemberResponse
+import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.request.SigninRequest
+import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.response.SigninResponse
+import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.request.SignupRequest
 import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.service.AuthService
 
 @RestController
@@ -17,21 +17,18 @@ import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.do
 class AuthController(
     val authService: AuthService,
 ) {
+
     @PostMapping("/signin")
-    fun signin(
-        @RequestBody request: SigninRequest
-    ): ResponseEntity<SigninResponse> {
-        return ResponseEntity
+    fun signin(@RequestBody request: SigninRequest): ResponseEntity<SigninResponse> =
+        ResponseEntity
             .status(HttpStatus.OK)
             .body(authService.signin(request))
-    }
+    
 
     @PostMapping("/signup")
-    fun signup(
-        @RequestBody request: SignupRequest
-    ): ResponseEntity<MemberResponse> {
-        return ResponseEntity
+    fun signup(@RequestBody request: SignupRequest): ResponseEntity<MemberResponse> =
+        ResponseEntity
             .status(HttpStatus.CREATED)
             .body(authService.signup(request))
-    }
+
 }
