--- src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/dto/MemberSimplifiedResponse.kt
@@ -1,8 +1,0 @@
-package spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto
-
-import java.util.UUID
-
-class MemberSimplifiedResponse(
-    val id: UUID,
-    val nickname: String
-)
