--- src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/dto/SigninResponse.kt
@@ -1,5 +1,0 @@
-package spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto
-
-data class SigninResponse(
-    val accessToken: String
-)
