--- src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/dto/UpdateProfileRequest.kt
@@ -1,5 +1,0 @@
-package spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto
-
-data class UpdateProfileRequest(
-    val nickname: String
-)
