+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/dto/request/UpdateProfileRequest.kt
@@ -1,0 +1,5 @@
+package spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.request
+
+data class UpdateProfileRequest(
+    val nickname: String
+)
