+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/dto/response/MemberSimplifiedResponse.kt
@@ -1,0 +1,8 @@
+package spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.response
+
+import java.util.UUID
+
+class MemberSimplifiedResponse(
+    val id: UUID,
+    val nickname: String
+)
