+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/domain/member/dto/response/SigninResponse.kt
@@ -1,0 +1,5 @@
+package spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.domain.member.dto.response
+
+data class SigninResponse(
+    val accessToken: String
+)
