+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/infra/security/MemberPrincipal.kt
@@ -10,7 +10,12 @@ data class MemberPrincipal(
     val name: String,
     val authorities: Collection<GrantedAuthority>
 ) {
-    constructor(id: String, email: String, roles: Set<String>) : this(
+
+    constructor(
+        id: String,
+        email: String,
+        roles: Set<String>
+    ) : this(
         UUID.fromString(id),
         email,
         roles.map { SimpleGrantedAuthority("ROLE_$it") })
