+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/infra/security/config/PasswordEncoderConfig.kt
@@ -7,6 +7,7 @@ import org.springframework.security.crypto.password.PasswordEncoder
 
 @Configuration
 class PasswordEncoderConfig {
+
     @Bean
     fun passwordEncoder(): PasswordEncoder {
         return BCryptPasswordEncoder()
