+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/infra/security/jwt/JwtAuthenticationToken.kt
@@ -9,6 +9,7 @@ class JwtAuthenticationToken(
     private val principal: MemberPrincipal,
     details: WebAuthenticationDetails,
 ) : AbstractAuthenticationToken(principal.authorities), Serializable {
+
     init {
         super.setAuthenticated(true)
         super.setDetails(details)
