+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/infra/security/jwt/JwtHelper.kt
@@ -7,7 +7,6 @@ import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.security.Keys
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
-import java.nio.charset.StandardCharsets
 import java.time.Duration
 import java.time.Instant
 import java.util.*
@@ -19,18 +18,27 @@ class JwtHelper(
     @Value("\${auth.jwt.accessTokenExpirationHour}") private val accessTokenExpirationHour: Int,
 ) {
 
-    fun validateToken(jwt: String): Result<Jws<Claims>> {
-        return kotlin.runCatching {
+    fun validateToken(jwt: String): Result<Jws<Claims>> =
+        kotlin.runCatching {
             Keys.hmacShaKeyFor(accessTokenSecret.toByteArray())
                 .let { Jwts.parser().verifyWith(it).build().parseSignedClaims(jwt) }
         }
-    }
 
-    fun generateAccessToken(subject: String, email: String, role: String): String {
-        return generateToken(subject, email, role, Duration.ofHours(accessTokenExpirationHour.toLong()))
-    }
 
-    private fun generateToken(subject: String, email: String, role: String, expirationPeriod: Duration): String {
+    fun generateAccessToken(
+        subject: String,
+        email: String,
+        role: String
+    ): String =
+        generateToken(subject, email, role, Duration.ofHours(accessTokenExpirationHour.toLong()))
+
+
+    private fun generateToken(
+        subject: String,
+        email: String,
+        role: String,
+        expirationPeriod: Duration
+    ): String {
         val claims = Jwts.claims().add(mapOf("email" to email, "role" to role)).build()
         val now = Instant.now()
         val key = Keys.hmacShaKeyFor(accessTokenSecret.toByteArray())
